requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# Web framework (for dashboard)
streamlit>=1.28.0
plotly>=5.15.0

# Visualization
folium>=0.14.0

# PostgreSQL database (optional)
psycopg2-binary>=2.9.0

# Environment and configuration
python-dotenv>=1.0.0

# API and HTTP requests (optional)
requests>=2.31.0

# Route optimization (optional - may have compatibility issues)
# ortools>=9.7.0
